+++ solvis/solution/solution_participation.py
@@ -6,7 +6,8 @@ from typing import TYPE_CHECKING, Iterable, Optional, cast
 
 import pandas as pd
 
-from solvis.filter import FilterSubsectionIds
+from solvis.filter import FilterParentFaultIds, FilterSubsectionIds
+from solvis.solution import named_fault
 
 from .typing import InversionSolutionProtocol
 
@@ -34,8 +35,9 @@ class SolutionParticipation:
         ```
 
     Methods:
-        section_participation_rates:  get rates for the specified fault sections.
-        fault_participation_rates: get rates for specificed faults.
+        section_participation_rates:  get rates for fault sections.
+        fault_participation_rates: get rates for parent faults.
+        named_fault_participation_rates: get rates for named faults
     """
 
     def __init__(self, solution: InversionSolutionProtocol):
@@ -58,9 +60,9 @@ class SolutionParticipation:
             rupture_ids: calculate participation using only these ruptures (aka `Conditional Participation`).
 
         Notes:
-         - Passing a non empty `subsection_ids` will not affect the rates, only the subsections for
+         - Passing a non empty `subsection_ids` does not affect the rates, only the subsections for
            which rates are returned.
-         - Passing a non empty `rupture_ids` will affect the rates, as only the specified ruptures
+         - Passing a non empty `rupture_ids` affects the rate calculation, as only the specified ruptures
            will be included in the sum.
            This is referred to as the `conditional participation rate` which might be used when you are
            only interested in the rates of e.g. ruptures in a particular magnitude range.
@@ -107,9 +109,9 @@ class SolutionParticipation:
             rupture_ids: calculate participation using only these ruptures (aka Conditional Participation).
 
         Notes:
-         - Passing `parent_fault_ids` will not affect the rate calculation, only the parent faults
+         - Passing `parent_fault_ids` does not affect the rate calculation, only the parent faults
            for which rates are returned.
-         - Passing `rupture_ids` will affect the rates, as only the specified ruptures
+         - Passing `rupture_ids` affects the rate calculation, as only the specified ruptures
            will be included in the sum.
            This is referred to as the `conditional participation rate` which might be used when you are
            only interested in e.g. the rates of ruptures in a particular magnitude range.
@@ -121,8 +123,6 @@ class SolutionParticipation:
             FilterSubsectionIds(self._solution).for_parent_fault_ids(parent_fault_ids) if parent_fault_ids else None
         )
 
-        rate_column = self._solution.model.rate_column_name()
-
         df0 = cast(pd.DataFrame, self._solution.model.rs_with_rupture_rates)
         if subsection_ids:
             df0 = df0[df0["section"].isin(subsection_ids)]
@@ -131,6 +131,7 @@ class SolutionParticipation:
             df0 = df0[df0["Rupture Index"].isin(rupture_ids)]
 
         df1 = df0.join(self._solution.solution_file.fault_sections[['ParentID']], on='section')
+        rate_column = self._solution.model.rate_column_name()
         result = (
             df1[["ParentID", "Rupture Index", rate_column]]
             .rename(columns={rate_column: 'participation_rate'})
@@ -141,3 +142,58 @@ class SolutionParticipation:
             .agg('sum')
         )
         return cast('DataFrame[dataframe_models.ParentFaultParticipationSchema]', result)
+
+    def named_fault_participation_rates(
+        self, named_fault_names: Optional[Iterable[str]] = None, rupture_ids: Optional[Iterable[int]] = None
+    ) -> 'DataFrame[dataframe_models.NamedFaultParticipationSchema]':
+        """Calculate the 'participation rate' for parent faults.
+
+        Participation rate for each named fault is the the sum of rupture rates for the
+        ruptures involving that fault.
+
+        Args:
+            named_fault_names: the list of named_fault_names to include.
+            rupture_ids: calculate participation using only these ruptures (aka Conditional Participation).
+
+        Notes:
+         - Passing `named_fault_names` does not affect the rate calculation, only the faults
+           for which rates are returned.
+         - Passing `rupture_ids` affects the rate calculation, as only the specified ruptures
+           will be included in the sum.
+           This is referred to as the `conditional participation rate` which might be used when you are
+           only interested in e.g. the rates of ruptures in a particular magnitude range.
+
+        Returns:
+            pd.DataFrame: a participation rates dataframe
+        """
+        parent_fault_ids = (
+            FilterParentFaultIds(self._solution).for_named_fault_names(named_fault_names) if named_fault_names else None
+        )
+
+        subsection_ids = (
+            FilterSubsectionIds(self._solution).for_parent_fault_ids(parent_fault_ids) if parent_fault_ids else None
+        )
+
+        df0 = cast(pd.DataFrame, self._solution.model.rs_with_rupture_rates)
+        if subsection_ids:
+            df0 = df0[df0["section"].isin(subsection_ids)]
+
+        if rupture_ids:
+            df0 = df0[df0["Rupture Index"].isin(rupture_ids)]
+
+        df1 = df0.join(self._solution.solution_file.fault_sections[['ParentID']], on='section')
+        df2 = df1.join(named_fault.named_fault_for_parent_ids_table(), 'ParentID', how='outer')
+        df2 = df2.drop(columns=['ParentID'])
+
+        rate_column = self._solution.model.rate_column_name()
+
+        result = (
+            df2[["named_fault_name", "Rupture Index", rate_column]]
+            .rename(columns={rate_column: 'participation_rate'})
+            .reset_index(drop=True)
+            .groupby(["named_fault_name", "Rupture Index"])
+            .agg('first')
+            .groupby("named_fault_name")
+            .agg('sum')
+        )
+        return cast('DataFrame[dataframe_models.NamedFaultParticipationSchema]', result)